+++ game/src/dlls/modgfx/109/109.c
@@ -36,7 +36,7 @@ u32 gDll6DEffectResourceData[sizeof(Dll6DEffectResourceView) / sizeof(u32)] = {
 
 void dll_6D_spawnEffect(GameObject* sourceObj, int variant, PartFxSpawnParams* spawnParams, u32 spawnFlags) {
     ModgfxSpawnPacket packet;
-    u8* resourceData = (u8*)(int)gDll6DEffectResourceData;
+    u8* resourceData = (u8*)gDll6DEffectResourceData;
     GameObject* context;
 
     packet.entries[0].layer = 0;
@@ -120,7 +120,7 @@ void dll_6D_spawnEffect(GameObject* sourceObj, int variant, PartFxSpawnParams* s
         }
     }
     (*gModgfxInterface)
-        ->spawnEffect(&packet, 0, 0xe, (u8*)(int)gDll6DEffectResourceData, 0xc,
+        ->spawnEffect(&packet, 0, 0xe, (u8*)gDll6DEffectResourceData, 0xc,
                       &resourceData[offsetof(Dll6DEffectResourceView, colors)], 0x34, 0);
 }
